style: Changed UI of the home screen
