+++ app/src/main/java/tk/talcharnes/unborify/MainActivity.java
@@ -40,8 +40,7 @@ public class MainActivity extends AppCompatActivity {
     private Toolbar toolbar;
     private NoSwipePager viewPager;
     private BottomBarAdapter pagerAdapter;
-    private Spinner spinner;
-    private int currentView = 0;
+    //private Spinner spinner;
 
     /**
      * Initializes basic initialization of components.
@@ -80,14 +79,14 @@ public class MainActivity extends AppCompatActivity {
         Uri uri = user.getPhotoUrl();
         String uriString = (uri != null) ? uri.toString() : "";
 
-        spinner = (Spinner) toolbar.findViewById(R.id.spinner);
+        //spinner = (Spinner) toolbar.findViewById(R.id.spinner);
         // Create an ArrayAdapter using the string array and a default spinner layout
-        ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(this,
-                R.array.spinner_list_item_array, android.R.layout.simple_spinner_item);
+        //ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(this,
+        //        R.array.spinner_list_item_array, android.R.layout.simple_spinner_item);
         // Specify the layout to use when the list of choices appears
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+        //adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
         // Apply the adapter to the spinner
-        spinner.setAdapter(adapter);
+        //spinner.setAdapter(adapter);
 
         setupViewPager();
 
@@ -95,23 +94,13 @@ public class MainActivity extends AppCompatActivity {
         bottomBar.setOnTabSelectListener(new OnTabSelectListener() {
             @Override
             public void onTabSelected(@IdRes int tabId) {
-                spinner.setVisibility(View.INVISIBLE);
-                if (tabId == R.id.tab_home) {
-                    spinner.setVisibility(View.VISIBLE);
-                    viewPager.setCurrentItem(0);
-                    currentView = 0;
-                } else if (tabId == R.id.tab_trending) {
-                    viewPager.setCurrentItem(1);
-                    currentView = 1;
-                } else if (tabId == R.id.tab_deals) {
-                    viewPager.setCurrentItem(2);
-                    currentView = 2;
-                } else if (tabId == R.id.tab_following) {
-                    viewPager.setCurrentItem(3);
-                    currentView = 3;
-                } else {
-                    viewPager.setCurrentItem(4);
-                    currentView = 4;
+                //spinner.setVisibility(View.INVISIBLE);
+                switch (tabId) {
+                    case R.id.tab_home: viewPager.setCurrentItem(0); break;
+                    case R.id.tab_trending: viewPager.setCurrentItem(1); break;
+                    case R.id.tab_deals: viewPager.setCurrentItem(2); break;
+                    case R.id.tab_following: viewPager.setCurrentItem(3); break;
+                    default: viewPager.setCurrentItem(4); break;
                 }
             }
         });
@@ -131,46 +120,5 @@ public class MainActivity extends AppCompatActivity {
         viewPager.setAdapter(pagerAdapter);
     }
 
-    /**
-     * This function handles hidden menu on the toolbar.
-     */
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        Log.d(TAG, "currentView: " + currentView);
-        if (currentView == 0) {
-            getMenuInflater().inflate(R.menu.main, menu);
-        }
-
-        return true;
-    }
-
-    /**
-     * This function handles the items clicked on the toolbar.
-     */
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        if (id == R.id.action_picture) {
-            Log.d(TAG, "Starting up the photo upload Activity.");
-            takePic();
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-
-    }
-
-    /**
-     * This function handles the action with the picture icon on the toolbar is clicked.
-     */
-    public void takePic() {
-        Intent intent = new Intent(this, PhotoUploadActivity.class);
-        startActivity(intent);
-    }
 
 }
