+++ app/src/main/res/layout/app_bar_main.xml
@@ -20,7 +20,7 @@
             android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
             app:popupTheme="@style/AppTheme.PopupOverlay">
 
-            <agency.tango.android.avatarview.views.AvatarView
+            <!--agency.tango.android.avatarview.views.AvatarView
                 android:id="@+id/avatarImage"
                 android:layout_width="32dp"
                 android:layout_height="32dp"
@@ -40,7 +40,7 @@
                 android:layout_marginEnd="12dp"
                 android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
                 app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-                tools:targetApi="lollipop" />
+                tools:targetApi="lollipop" /-->
 
         </android.support.v7.widget.Toolbar>
