+++ app/src/main/res/layout/photo_card_view.xml
@@ -33,8 +33,8 @@
             android:id="@+id/photoImageView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginTop="55dp"
-            android:scaleType="fitXY"
+            android:layout_marginBottom="50dp"
+            android:scaleType="centerCrop"
             android:src="@mipmap/ic_launcher" />
 
         <android.support.design.widget.FloatingActionButton
@@ -55,10 +55,27 @@
             android:layout_centerInParent="true"
             android:visibility="visible" />
 
+        <com.iarcuschin.simpleratingbar.SimpleRatingBar
+            android:id="@+id/ratingbar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="#80000000"
+            android:layout_above="@+id/card_bottom_layout"
+            android:contentDescription="@string/rating_bar"
+            app:srb_borderColor="#FFF"
+            app:srb_fillColor="@color/colorPrimary"
+            app:srb_isIndicator="true"
+            app:srb_numberOfStars="5"
+            app:srb_rating="0"
+            app:srb_starSize="36dp"
+            app:srb_stepSize="0.1" />
+
         <LinearLayout
+            android:id="@+id/card_bottom_layout"
             android:layout_width="match_parent"
             android:layout_height="50dp"
-            android:layout_gravity="top"
+            android:layout_gravity="bottom"
+            android:layout_alignParentBottom="true"
             android:gravity="center"
             android:orientation="horizontal"
             android:paddingEnd="15dp"
@@ -70,13 +87,13 @@
                 android:layout_height="42dp"
                 android:gravity="start"
                 android:scaleType="centerCrop"
-                app:av_border_color="@color/colorPrimary"
-                app:av_border_width="4dp"
+                app:av_border_color="@color/white"
+                app:av_border_width="1dp"
                 app:av_text_size_percentage="75" />
 
             <LinearLayout
                 android:layout_width="0dp"
-                android:layout_height="55dp"
+                android:layout_height="50dp"
                 android:layout_gravity="bottom"
                 android:layout_weight="2"
                 android:gravity="center|start"
@@ -85,47 +102,25 @@
                 android:paddingStart="20dp">
 
                 <TextView
-                    android:id="@+id/occasion_textview"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:textAlignment="center"
-                    android:textColor="@color/colorPrimaryDark"
-                    android:textSize="13sp"
-                    android:textStyle="bold"
-                    tools:text="@string/no_deals_available_string" />
-
-                <TextView
                     android:id="@+id/uploadederNameTxt"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:textColor="@color/colorPrimaryDark"
-                    android:textSize="14sp"
+                    android:textSize="15sp"
+                    android:textStyle="bold"
+                    tools:text="Name" />
+
+                <TextView
+                    android:id="@+id/occasion_textview"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textAlignment="center"
+                    android:textColor="@color/colorPrimary"
+                    android:textSize="13sp"
                     android:textStyle="normal"
-                    tools:text="@string/no_deals_available_string" />
+                    tools:text="Occasion" />
+
             </LinearLayout>
-
-        </LinearLayout>
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
-            android:layout_alignParentBottom="true"
-            android:layout_centerHorizontal="true"
-            android:background="#80000000">
-
-            <com.iarcuschin.simpleratingbar.SimpleRatingBar
-                android:id="@+id/ratingbar"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_centerInParent="true"
-                android:contentDescription="@string/rating_bar"
-                app:srb_borderColor="#FFF"
-                app:srb_fillColor="@color/colorPrimary"
-                app:srb_isIndicator="true"
-                app:srb_numberOfStars="5"
-                app:srb_rating="0"
-                app:srb_starSize="36dp"
-                app:srb_stepSize="0.1" />
 
             <ImageButton
                 android:id="@+id/photo_card_options"
@@ -134,11 +129,9 @@
                 android:layout_gravity="center|end"
                 android:background="@color/white"
                 android:contentDescription="@string/picture_options"
-                android:layout_alignParentEnd="true"
-                android:layout_centerVertical="true"
                 android:src="@drawable/more_options" />
 
-        </RelativeLayout>
+        </LinearLayout>
 
     </RelativeLayout>
