+++ app/src/main/res/values/colors.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimaryLight">#4f648b</color>
-    <color name="colorPrimary">#223A5E</color>
-    <color name="colorPrimaryDark">#001434</color>
+    <color name="colorPrimaryLight">#B3E5FC</color>
+    <color name="colorPrimary">#03A9F4</color>
+    <color name="colorPrimaryDark">#0288D1</color>
     <color name="primary_light">#B3E5FC</color>
     <color name="colorAccent">#FF4081</color>
     <color name="primary_text">#212121</color>
@@ -13,6 +13,7 @@
     <color name="black_overlay">#66000000</color>
 
     <color name="white">#FFFFFF</color>
+    <color name="black">#000000</color>
 
     <color name="menu_item_color">#D9000000</color>
